프론트와 호환되게 변수 수정

+++ Barotruck/src/main/java/com/example/project/domain/order/dto/OrderResponse.java
@@ -79,8 +79,8 @@ public class OrderResponse {
     private Long endNbhId;
 
     // 시스템 지표
-    private Long distance;
-    private Long duration;
+    private BigDecimal distance;
+    private BigDecimal duration;
 
     private UserSummary user;
     private CancellationSummary cancellation;
